Physics calculation for new net engine, the only things left to do is network and option in FeatureList
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3536 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/server/CServer_Parse.cpp
@@ -891,7 +891,7 @@ void GameServer::ParseConnect(NetworkSocket tSocket, CBytestream *bs) {
 
 	// Get user info
 	int numworms = bs->readInt(1);
-	numworms = CLAMP(numworms, 0, MAX_PLAYERS);
+	numworms = CLAMP(numworms, 0, (int)MAX_PLAYERS);
 	
 	// If we ignored this challenge verification, there could be double connections
